refactor(esbabbler): extract room list item settings button
